+++ types/linkedom.d.ts
@@ -1,0 +1,3 @@
+declare module "linkedom" {
+    export function parseHTML(html: any): Window & typeof globalThis;
+}